pyproject.toml
[tool.poetry]
name = "qr-picturizer"
version = "0.1.0"
description = "Manages the dependencies of qr-picturizer"
authors = ["Niko Kolaxidis <nikolaos.kolaxidis@heigit.org>"]
license = "GPL-3.0-or-later"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
qrcode = "^7.4.2"
pillow = "^10.1.0"
pyyaml = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
